1. (9585) All the Fan Relay's will work based on the conditioning mode selected on user intent.
2. (9586) All the Analog out signals are resetting based on the conditioning mode selected
3. Operating mode points is showing  according the conditioning
4. (9549) Now we are actually conditioning when zone in auto away
5. Removed non developed profile selection option from select module screen

Related work items: #9549, #9585, #9586, #9616

+++ app/src/main/java/a75f/io/renatus/hyperstat/cpu/HyperstatCpuViewModel.kt
@@ -424,7 +424,7 @@ private fun offsetSpinnerValues(
 
     val nums = arrayOfNulls<String>(count)
     for (nNum in 0 until count) {
-        var rawValue = (nNum + offsetFromZeroCount).toFloat() * inc
+        var rawValue = (nNum + offsetFromZeroCount).toDouble() * inc
         if (displayAsInt) {
             nums[nNum] = String.format("%.1f", rawValue) + suffix
         } else {
